backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
fastapi==0.129.2
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
PyJWT==2.8.0
bcrypt==4.0.1
passlib==1.7.4
email-validator==2.1.0
python-multipart==0.0.9
firebase-admin==6.5.0
supabase==2.3.4
python-dotenv==1.0.0
httpx>=0.24,<0.26
gotrue<2.9.0
